Copying changes to a new branch...

+++ Mona.SaaS/Mona.SaaS.Setup/basic-deploy.sh
@@ -4,18 +4,22 @@ mona_version="0.1-prerelease"
 
 exec 3>&2 # Grabbing a reliable stderr handle...
 
-usage() { printf "\nUsage: $0 <-n deployment-name> <-r deployment-region> [-d display-name] [-g resource-group] [-l ui-language] [-s subscription-id]\n"; }
+usage() { printf "\nUsage: $0 <-n deployment-name> <-r deployment-region> [-a app-service-plan-id] [-d display-name] [-g resource-group] [-l ui-language] [-s subscription-id] [-h] [-p]\n"; }
 
 check_az() {
     exec 3>&2
 
     az version >/dev/null 2>&1
 
+    lp=$1
+
     # TODO: Should we be more specific about which version of az is required?
 
     if [[ $? -ne 0 ]]; then 
-        printf "\nPlease install the Azure CLI before continuing. See [https://docs.microsoft.com/cli/azure/install-azure-cli] for more information.\n" >&3
+        echo "$lp ❌   Please install the Azure CLI before continuing. See [https://docs.microsoft.com/cli/azure/install-azure-cli] for more information."
         return 1
+    else
+        echo "$lp ✔   Azure CLI installed."
     fi
 }
 
@@ -24,53 +28,81 @@ check_dotnet() {
 
     dotnet --version >/dev/null 2>&1
 
+    lp=$1
+
    # TODO: Should we be more specific about which version of dotnet is required?
 
     if [[ $? -ne 0 ]]; then 
-        printf "\nPlease install .NET before continuing. See [https://dotnet.microsoft.com/download] for more information.\n" >&3
+        echo "$lp ❌   Please install .NET before continuing. See [https://dotnet.microsoft.com/download] for more information."
         return 1
+    else
+        echo "$lp ✔   .NET installed."
     fi
 }
 
 check_prereqs() {
-    printf "Checking Mona setup prerequisites..."; echo
+    lp=$1
 
-    check_az;         if [[ $? -ne 0 ]]; then prereq_check_failed=1; fi; 
-    check_dotnet;     if [[ $? -ne 0 ]]; then prereq_check_failed=1; fi;
+    echo "$lp Checking Mona setup prerequisites...";
+
+    check_az "$lp";         if [[ $? -ne 0 ]]; then prereq_check_failed=1; fi; 
+    check_dotnet "$lp";     if [[ $? -ne 0 ]]; then prereq_check_failed=1; fi;
 
     if [[ -z $prereq_check_failed ]]; then
-        printf "\nAll Mona setup prerequisites installed.\n"
+        echo "$lp ✔   All Mona setup prerequisites installed."
     else
         return 1
     fi
 }
 
-build_mona() {
-    mona_sln_path="../Mona.SaaS.sln"
+check_app_service_plan() {
+    lp=$1
+    plan_id=$2
 
-    printf "\nBuilding Mona...\n\n";
+    plan_name=$(az appservice plan show --ids "$plan_id" --output "tsv" --query "name");
 
-    dotnet build "$mona_sln_path"
-
-    if [[ $? -ne 0 ]]; then
-        printf "\nMona build has failed. Please correct issues then try again.\n" >&3
-        exit 1
+    if [[ -z $plan_name ]]; then
+        echo "$lp ❌   App service plan [$plan_id] not found."
+        return 1
+    else
+        echo "$lp ✔   Will deploy Mona web app to app service plan [$plan_name]."
     fi
+}
 
-    printf "\nRunning Mona tests...\n\n"
+check_deployment_region() {
+    lp=$1
+    region=$2
 
-    dotnet test "$mona_sln_path"
+    region_display_name=$(az account list-locations -o tsv --query "[?name=='$region'].displayName")
 
-    if [[ $? -ne 0 ]]; then
-        printf "\nOne or more Mona tests has failed. Please correct issues then try again.\n" >&3
-        exit 1
+    if [[ -z $region_display_name ]]; then
+        echo "$lp ❌   [$region] is not a valid Azure region. For a full list of Azure regions, run 'az account list-locations -o table'."
+        return 1
+    else
+        echo "$lp ✔   [$region] is a valid Azure region ($region_display_name)."
+    fi
+}
+
+check_language() {
+    lp=$1
+    language=$2
+    supported_languages=("en" "es")
+
+    if [[ " ${supported_languages[*]} " == *"$language"* ]]; then
+        echo "$lp ✔   [$language] language is supported."
+    else
+        echo "$lp ❌   [$language] language is not supported."
+        return 1
     fi
 }
 
 language="en" # Default UI language is English ("en"). Can be overridden using [-l] flag below.
 
-while getopts "d:g:l:n:r:s:" opt; do
+while getopts "a:d:g:l:n:r:s:hp" opt; do
     case $opt in
+        a)
+            app_service_plan_id=$OPTARG
+        ;;
         d)
             display_name=$OPTARG
         ;;
@@ -89,6 +121,12 @@ while getopts "d:g:l:n:r:s:" opt; do
         s)
             subscription_id=$OPTARG
         ;;
+        h)
+            no_splash=1
+        ;;
+        p)
+            no_publish=1
+        ;;
         \?)
             usage
             exit 1
@@ -96,26 +134,46 @@ while getopts "d:g:l:n:r:s:" opt; do
     esac
 done
 
-# Validate parameters.
+# Check for missing parameters.
+# Set default resource group name and deployment display name.
 
 [[ -z $deployment_name || -z $deployment_region ]] && { usage; exit 1; }
 [[ -z $resource_group_name ]] && resource_group_name="mona-$deployment_name";
 [[ -z $display_name ]] && display_name="$deployment_name";
 
+lp=$(printf '[%s%*s]>' "$deployment_name" "$((13-${#deployment_name}))" "");
+
+echo "$lp Setting up Mona SaaS in your Azure environment...";
+
+# Show Mona setup splash screen...
+
+[[ -z $no_splash ]] && cat ./splash.txt && echo;
+
+echo "$lp This Mona deployment's name is [$deployment_name]."
+
 # Check setup pre-reqs.
 
-check_prereqs
+check_prereqs "$lp"
 
 if [[ $? -ne 0 ]]; then
-    printf "\nPlease install all Mona setup prerequisites then try again.\n" >&3
+    echo "$lp ❌   Please install all Mona setup prerequisites then try again. Setup failed."
     exit 1
 fi
 
-# Build/test Mona locally.
+# Check parameter values.
 
-build_mona
+check_deployment_region "$lp" "$deployment_region"; if [[ $? -ne 0 ]]; then param_valid_failed=1; fi;
+check_language "$lp" "$language"; if [[ $? -ne 0 ]]; then param_valid_failed=1; fi;
 
-if [[ $? -ne 0 ]]; then exit 1; fi; # Mona build has failed.
+if [[ -n $app_service_plan_id ]]; then
+    check_app_service_plan "$lp" "$app_service_plan_id"
+    if [[ $? -ne 0 ]]; then exit 1; fi;
+fi
+
+if [[ -n $param_valid_failed ]]; then 
+    echo "$lp ❌   Parameter validation failed. Please review then try again. Setup failed."
+    exit 1
+fi
 
 # Ensure user is logged in to Azure.
 # Get current user object ID (oid).
@@ -129,16 +187,24 @@ done
 
 if [[ -n $subscription_id ]]; then
     az account set --subscription $subscription_id
+    [[ $? -ne 0 ]] && echo "$lp ❌   Azure subscription [$subscription_id] not found. Setup failed" && exit 1;
 fi
 
 # Create the resource group if it doesn't already exist.
 # If it already exists confirm that it's empty.
 
-if [[ $(az group exists --resource-group "$resource_group_name" --output tsv) -eq false ]]; then
-    printf "\nCreating resource group [$resource_group_name]...\n"
-    az group create --location "$deployment_region" --name "$resource_group_name" >/dev/null
+if [[ $(az group exists --resource-group "$resource_group_name") -eq false ]]; then
+    echo "$lp Creating resource group [$resource_group_name]..."
+    az group create --location "$deployment_region" --name "$resource_group_name"
+
+    if [[ $? -eq 0 ]]; then 
+        echo "$lp ✔   Resource group [$resource_group_name] created."
+    else
+        echo "$lp ❌   Unable to create resource group [$resource_group_name]. See above output for details. Setup failed."
+        exit 1
+    fi
 elif [[ -n $(az resource list --resource-group "$resource_group_name" --output tsv) ]]; then
-    printf "\nMona must be deployed into an empty resource group. Resource group [$resource_group_name] contains resources. Setup failed.\n" >2
+    echo "$lp ❌   Mona must be deployed into an empty resource group. Resource group [$resource_group_name] contains resources. Setup failed."
     exit 1
 fi
 
@@ -150,7 +216,7 @@ current_user_tid=$(az account show --query tenantId --output tsv);
 aad_app_name="$display_name"
 aad_app_secret=$(openssl rand -base64 64)
 
-printf "\nCreating Azure Active Directory (AAD) app registration [$aad_app_name]...\n"
+echo "$lp Creating Azure Active Directory (AAD) app registration [$aad_app_name]..."
 
 aad_app_id=$(az ad app create \
     --display-name "$aad_app_name" \
@@ -162,14 +228,14 @@ aad_app_id=$(az ad app create \
     --query appId \
     --output tsv);
 
-printf "\nAAD app [$aad_app_name ($aad_app_id)] successfully registered with AAD tenant [$current_user_tid].\n"
-printf "\nCreating app service principal. This might take a while...\n"
+echo "$lp ✔   AAD app [$aad_app_name ($aad_app_id)] successfully registered with AAD tenant [$current_user_tid]."
+echo "$lp Creating app service principal. This might take a while..."
 
 sleep 30 # Give AAD a chance to catch up...
 
 aad_sp_id=$(az ad sp create --id "$aad_app_id" --query objectId --output tsv 2>/dev/null);
 
-printf "\nGranting AAD app [$aad_app_name] service principal [$aad_sp_id] contributor access to resource group [$resource_group_name]...\n\n"
+echo "$lp Granting AAD app [$aad_app_name] service principal [$aad_sp_id] contributor access to resource group [$resource_group_name]..."
 
 sleep 30 # Give AAD a chance to catch up...
 
@@ -178,7 +244,7 @@ az role assignment create \
     --assignee "$aad_sp_id" \
     --resource-group "$resource_group_name"
 
-printf "\nTagging resource group [$resource_group_name]...\n"
+echo "$lp Tagging resource group [$resource_group_name]..."
 
 az group update \
     --name "$resource_group_name" \
@@ -190,7 +256,7 @@ az group update \
 
 # Deploy the ARM template.
 
-printf "\nDeploying Mona to subscription [$subscription_id] resource group [$resource_group_name]. This might take a while...\n";
+echo "$lp Deploying Mona to subscription [$subscription_id] resource group [$resource_group_name]. This might take a while...";
 
 az_deployment_name="mona-deploy-$deployment_name"
 
@@ -203,7 +269,11 @@ az group deployment create \
         aadTenantId="$current_user_tid" \
         aadClientId="$aad_app_id" \
         aadClientSecret="$aad_app_secret" \
-        language="$language"; echo;
+        language="$language" \
+        appServicePlanId="$app_service_plan_id"
+
+[[ $? -eq 0 ]] && echo "$lp ✔   Mona resources successfully deployed [$az_deployment_name] to resource group [$resource_group_name].";
+[[ $? -ne 0 ]] && echo "$lp ❌   Mona resource group [$resource_group_name] deployment [$az_deployment_name] has failed. Aborting setup..." && exit 1;
         
 # Get ARM deployment output variables.
 
@@ -220,28 +290,28 @@ app_insights_conn_str=$(az deployment group show --resource-group "$resource_gro
 
 # Configure Mona.
 
-printf "\nConfiguring Mona settings...\n\n";
+echo "$lp Configuring Mona settings...";
 
-az appconfig kv set --name "$app_config_name" --key "Deployment:MonaVersion" --yes                                      --value "$mona_version"; echo;
-az appconfig kv set --name "$app_config_name" --key "Deployment:AppInsightsConnectionString" --yes                      --value "$app_insights_conn_str"; echo;
-az appconfig kv set --name "$app_config_name" --key "Deployment:AzureResourceGroupName" --yes                           --value "$resource_group_name"; echo;
-az appconfig kv set --name "$app_config_name" --key "Deployment:AzureSubscriptionId" --yes                              --value "$subscription_id"; echo;
-az appconfig kv set --name "$app_config_name" --key "Deployment:IsTestModeEnabled" --yes                                --value "true"; echo;
-az appconfig kv set --name "$app_config_name" --key "Deployment:Name" --yes                                             --value "$deployment_name"; echo;
-az appconfig kv set --name "$app_config_name" --key "Identity:AdminIdentity:AadTenantId" --yes                          --value "$current_user_tid"; echo;
-az appconfig kv set --name "$app_config_name" --key "Identity:AdminIdentity:AadUserId" --yes                            --value "$current_user_oid"; echo;
-az appconfig kv set --name "$app_config_name" --key "Identity:AppIdentity:AadClientId" --yes                            --value "$aad_app_id"; echo;
-az appconfig kv set --name "$app_config_name" --key "Identity:AppIdentity:AadClientSecret" --yes                        --value "$aad_app_secret" >/dev/null # Sensitive
-az appconfig kv set --name "$app_config_name" --key "Identity:AppIdentity:AadTenantId" --yes                            --value "$current_user_tid"; echo;
-az appconfig kv set --name "$app_config_name" --key "Offer:IsSetupComplete" --yes                                       --value "false"; echo;
-az appconfig kv set --name "$app_config_name" --key "Subscriptions:Events:EventGrid:TopicEndpoint" --yes                --value "$event_grid_topic_endpoint"; echo; 
-az appconfig kv set --name "$app_config_name" --key "Subscriptions:Events:EventGrid:TopicKey" --yes                     --value "$event_grid_topic_key" >/dev/null # Sensitive
-az appconfig kv set --name "$app_config_name" --key "Subscriptions:Repository:BlobStorage:ConnectionString" --yes       --value "$blob_conn_str" >/dev/null # Sensitive
-az appconfig kv set --name "$app_config_name" --key "Subscriptions:Repository:BlobStorage:ContainerName" --yes          --value "$blob_sub_container_name"; echo;
+az appconfig kv set --name "$app_config_name" --key "Deployment:MonaVersion" --yes                                      --value "$mona_version";                
+az appconfig kv set --name "$app_config_name" --key "Deployment:AppInsightsConnectionString" --yes                      --value "$app_insights_conn_str";       
+az appconfig kv set --name "$app_config_name" --key "Deployment:AzureResourceGroupName" --yes                           --value "$resource_group_name";         
+az appconfig kv set --name "$app_config_name" --key "Deployment:AzureSubscriptionId" --yes                              --value "$subscription_id";             
+az appconfig kv set --name "$app_config_name" --key "Deployment:IsTestModeEnabled" --yes                                --value "true";                         
+az appconfig kv set --name "$app_config_name" --key "Deployment:Name" --yes                                             --value "$deployment_name";             
+az appconfig kv set --name "$app_config_name" --key "Identity:AdminIdentity:AadTenantId" --yes                          --value "$current_user_tid";
+az appconfig kv set --name "$app_config_name" --key "Identity:AdminIdentity:AadUserId" --yes                            --value "$current_user_oid";
+az appconfig kv set --name "$app_config_name" --key "Identity:AppIdentity:AadClientId" --yes                            --value "$aad_app_id";
+az appconfig kv set --name "$app_config_name" --key "Identity:AppIdentity:AadClientSecret" --yes                        --value "$aad_app_secret" >/dev/null
+az appconfig kv set --name "$app_config_name" --key "Identity:AppIdentity:AadTenantId" --yes                            --value "$current_user_tid";
+az appconfig kv set --name "$app_config_name" --key "Offer:IsSetupComplete" --yes                                       --value "false";
+az appconfig kv set --name "$app_config_name" --key "Subscriptions:Events:EventGrid:TopicEndpoint" --yes                --value "$event_grid_topic_endpoint"; 
+az appconfig kv set --name "$app_config_name" --key "Subscriptions:Events:EventGrid:TopicKey" --yes                     --value "$event_grid_topic_key" >/dev/null
+az appconfig kv set --name "$app_config_name" --key "Subscriptions:Repository:BlobStorage:ConnectionString" --yes       --value "$blob_conn_str" >/dev/null
+az appconfig kv set --name "$app_config_name" --key "Subscriptions:Repository:BlobStorage:ContainerName" --yes          --value "$blob_sub_container_name"; 
 
 # Configure AD app reply and ID URLs.
 
-printf "Completing Mona configuration...\n"
+echo "$lp Completing Mona configuration..."
 
 az ad app update \
     --id "$aad_app_id" \
@@ -254,66 +324,52 @@ az webapp config appsettings set \
     --name "$web_app_name" \
     --settings APP_CONFIG_SERVICE_CONNECTION_STRING="$app_config_connection_string" >/dev/null # Sensitive
 
-# Deploy Mona web application...
+if [[ -z $no_publish ]]; then
+    # Deploy Mona web application...
 
-printf "\nPackaging Mona web app for deployment to [$web_app_name]...\n\n"
+    echo "$lp Packaging Mona web app for deployment to [$web_app_name]..."
 
-dotnet publish -c Release -o ./topublish ../Mona.SaaS.Web/Mona.SaaS.Web.csproj
+    dotnet publish -c Release -o ./topublish ../Mona.SaaS.Web/Mona.SaaS.Web.csproj
 
-printf "\nZipping Mona web app deployment package...\n"
+    echo "$lp Zipping Mona web app deployment package..."
 
-cd ./topublish
-zip -r ../topublish.zip . >/dev/null
-cd ..
+    cd ./topublish
+    zip -r ../topublish.zip . >/dev/null
+    cd ..
 
-printf "\nDeploying Mona web app to [$web_app_name]...\n"
+    echo "$lp Deploying Mona web app to [$web_app_name]..."
 
-az webapp deployment source config-zip -g "$resource_group_name" -n "$web_app_name" --src ./topublish.zip
+    az webapp deployment source config-zip -g "$resource_group_name" -n "$web_app_name" --src ./topublish.zip
 
-# And scene...
+    # And scene...
 
-printf "Cleaning up...\n"
+    echo "$lp Cleaning up..."
 
-rm -rf ./topublish >/dev/null
-rm -rf ./topublish.zip >/dev/null
+    rm -rf ./topublish >/dev/null
+    rm -rf ./topublish.zip >/dev/null
+fi
 
-printf "\nMona Deployment Summary\n"
-printf "==============================\n"
-printf "Deployment Name                     [$deployment_name]\n"
-printf "Deployment Version                  [$mona_version]\n"
-printf "Deployed to Azure Subscription      [$subscription_id]\n"
-printf "Deployed to Resource Group          [$resource_group_name]\n"
-printf "Deployment AAD Client ID            [$aad_app_id]\n"
-printf "Deployment AAD Tenant ID            [$current_user_tid]\n"
-printf "Landing Page URL                    [$web_app_base_url/]\n"
-printf "Landing Page URL (Testing)          [$web_app_base_url/test]\n"
-printf "Webhook URL                         [$web_app_base_url/webhook]\n"
-printf "Webhook URL (Testing)               [$web_app_base_url/webhook/test]\n"
-printf "Admin Center URL                    [$web_app_base_url/admin]\n"
+printf "\n$lp Mona Deployment Summary\n"
+echo
+printf "$lp Deployment Name                     [$deployment_name]\n"
+printf "$lp Deployment Version                  [$mona_version]\n"
+printf "$lp Deployed to Azure Subscription      [$subscription_id]\n"
+printf "$lp Deployed to Resource Group          [$resource_group_name]\n"
+printf "$lp Deployment AAD Client ID            [$aad_app_id]\n"
+printf "$lp Deployment AAD Tenant ID            [$current_user_tid]\n"
 
-printf "\nMona deployment complete.\n"
-printf "\n==> Please visit [ $web_app_base_url/setup ] to complete your setup and begin transacting with Microsoft!\n"
+if [[ -z $no_publish ]]; then
+    printf "$lp Landing Page URL                    [$web_app_base_url/]\n"
+    printf "$lp Landing Page URL (Testing)          [$web_app_base_url/test]\n"
+    printf "$lp Webhook URL                         [$web_app_base_url/webhook]\n"
+    printf "$lp Webhook URL (Testing)               [$web_app_base_url/webhook/test]\n"
+    printf "$lp Admin Center URL                    [$web_app_base_url/admin]\n"
+fi
 
+echo
+echo "$lp ✔   Mona deployment complete."
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+if [[ -z $no_publish ]]; then
+    echo
+    echo "$lp ▶   Please visit [ $web_app_base_url/setup ] to complete your setup and begin transacting with Microsoft!"
+fi
